@@ -1,9 +1,8 @@
+from importlib.metadata import version
 from datetime import datetime
-from pkg_resources import get_distribution
 import os
 import shutil
 
-version = get_distribution('aws_secretsmanager_caching').version
 project = u'AWS Secrets Manager Python Caching Client'
 
 # If you use autosummary, this ensures that any stale autogenerated files are
@@ -32,7 +31,7 @@ master_doc = 'index'  # The master toctree document.
 copyright = u'%s, Amazon.com' % datetime.now().year
 
 # The full version, including alpha/beta/rc tags.
-release = version
+release = version('aws_secretsmanager_caching')
 
 # List of directories, relative to source directory, that shouldn't be searched
 # for source files.
